Extract AuthGate
