+++ lib/auth_gate.dart
@@ -1,0 +1,28 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:flutterfire_ui/auth.dart';
+
+import 'pages/home_page.dart';
+
+class AuthGate extends StatelessWidget {
+  const AuthGate({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder<User?>(
+      stream: FirebaseAuth.instance.authStateChanges(),
+      builder: (context, snapshot) {
+        // The user is not signed in.
+        if (!snapshot.hasData) {
+          return const SignInScreen(
+            providerConfigs: [
+              EmailProviderConfiguration(),
+            ],
+          );
+        }
+
+        return const HomePage();
+      },
+    );
+  }
+}
